Move creation of rewrite recipe from string into RewriteRecipeLoader
- OpenRewriteDeclarativeRecipeAdapter uses RewriteRecipeLoader to be close to OpenRewriteNamedRecipeAdapter
- RewriteRecipeLoader offers method to create rewrite Recipe from String
- Adds tests
- Add comments
- Remove integration test covered by unittest
