+++ components/sbm-core/src/main/java/org/springframework/sbm/engine/recipe/RewriteRecipeRunner.java
@@ -27,7 +27,7 @@ import org.springframework.stereotype.Component;
 import java.util.List;
 
 @Component
-public class OpenRewriteRecipeRunner {
+public class RewriteRecipeRunner {
     @Autowired
     private RewriteMigrationResultMerger resultMerger;
